backend/pyproject.toml
[project]
name = "harmony-chat-api"
version = "0.1.0"
description = "Implementation of harmony-chat backend"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "scalar-fastapi>=1.6.2",

    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.12.0",
    "typing-inspection>=0.4.2",
    "annotated-doc>=0.0.4",
    "structlog>=25.5.0",

    "aioboto3>=15.5.0",
    "aiofiles>=25.1.0",
    "redis>=7.2.0",
    "cent>=5.2.0",
    "sqlalchemy>=2.0.46",
    "asyncpg>=0.31.0",

    "python-ulid>=3.1.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
]

[tool.setuptools.packages.find]
where = ["src"]